refactor(geodesy): rename rectangular_extent module
Rename rectangular_extent module to rectangular_region.
