+++ src/goesdl/geodesy/rectangular_region.py
